pyproject.toml
[tool.poetry]
name = "kohocarpet"
version = "0.1.0"
description = "Kohonens map in python"
authors = ["Jiri Podivin <jpodivin@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.0"
matplotlib = "^3.2.2"
Pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kohocarpet = 'kohocarpet.koho_carpet:main'
